Create sum = sum + k;
why does k = 103  instead of printing out all the values of k :   #include <iostream>
using namespace std;

int main()
{
 int k;
  
  for (k = 1; k <= 10; k = k + 1)
   cout << "k = " << k << endl;
// make these symbolic constants so that if we wish to change them
  // we can do it quickly up here rather than muck around in the code
  int const LOW_VAL = 15;
  int const HIGH_VAL = 101;
  

  int sum;  // holds the answer
   // loop counter

  sum = 0;  // start it off at 0 - we will add to it
  // loop, adding the ODD numbers
  for ( k = LOW_VAL; k <= HIGH_VAL; k = k + 2 )
   sum = sum + k;
  cout <<" k = " << k << endl;
  cout << "The sum of the odd numbers between " << LOW_VAL
   << " and " << HIGH_VAL << ", inclusive, is " << sum << endl << endl;
  
View other drafts


















The reason why k = 103 instead of printing out all the values of k is because the second for loop overwrites the value of k. The second for loop initializes k to 15, and then loops until k is greater than or equal to 101. At the end of the loop, k is equal to 103.

To fix this, you can move the line cout << "k = " << k << endl; to the inside of the second for loop, before the line sum = sum + k;. This will print the value of k each time the loop iterates.

Here is the modified code:

C++
#include <iostream>
using namespace std;

int main()
{
 int k;
  
  for (k = 1; k <= 10; k = k + 1)
   cout << "k = " << k << endl;
// make these symbolic constants so that if we wish to change them
  // we can do it quickly up here rather than muck around in the code
  int const LOW_VAL = 15;
  int const HIGH_VAL = 101;
  

  int sum;  // holds the answer
   // loop counter

  sum = 0;  // start it off at 0 - we will add to it
  // loop, adding the ODD numbers
  for ( k = LOW_VAL; k <= HIGH_VAL; k = k + 2 )
  {
   cout << "k = " << k << endl;
   sum = sum + k;
   }
  cout << "The sum of the odd numbers between " << LOW_VAL
   << " and " << HIGH_VAL << ", inclusive, is " << sum << endl << endl;
  
}