Create add all the odd numbers from 15 to 1001
The first for loop can omit brackets because it has only one statement inside of it. In C++, curly braces are used to group multiple statements together. If there is only one statement, then the curly braces are not necessary.

The second for loop has two statements inside of it, so the curly braces are necessary. This is because the compiler needs to know which statements belong to the for loop and which ones don't. Without the curly braces, the compiler would not be able to determine which statements should be executed each time the loop iterates.

Here is a diagram that illustrates the difference:

First for loop:

for (k = 1; k <= 10; k = k + 1)
  cout << "k = " << k << endl;

Second for loop:

for ( k = LOW_VAL; k <= HIGH_VAL; k = k + 2 )
{
  cout << "k = " << k << endl;
  sum = sum + k;
}
In the first for loop, the cout statement is the only statement inside of the loop. Therefore, the curly braces are not necessary.

In the second for loop, there are two statements inside of the loop: the cout statement and the sum statement. Therefore, the curly braces are necessary to group the statements together and tell the compiler which statements belong to the for loop.

It is generally considered good practice to use curly braces for all for loops, even if there is only one statement inside of the loop. This makes the code more readable and easier to maintain.




o fix this, you can move the line cout << "k = " << k << endl; to the inside of the second for loop, before the line sum = sum + k;. This will print the value of k each time the loop iterates.

Here is the modified code:

C++
#include <iostream>
using namespace std;

int main()
{