Create math with user input
Let's look at the calcPCT function again.

#include <iostream>
using namespace std;

float calcPCT(float total, float lostpoints)
{
  float pct;
  pct = (total - lostpoints)/total;
  return pct;
}

int main()
{
 float pct = calcPCT(100, 10);
 cout << pct << endl;
}
Now that we know how to get input from the user, let's make this program a little more interactive.  All of the user interaction code will be added to the main function.  This is to create a clear separation between the user interaction code and the calculation.  It makes it easier to test, debug and make updates when we keep the user interaction code separate and distinct from the business logic of a program.  You could put the user interaction code in its own separate function if you prefer, but we will put it in main for now to keep things simple.

Here are the basic steps to adding user interaction:

Declare one or more variables that will eventually store the input
Use cout to display instructions to the user
Use cin to get the user's input and store it in the variables you declared previously
Repeat these steps for as many inputs as needed.  Once you get the input, you can pass it along to the calcPCT function by including the variable(s) as arguments when you call the function.

#include <iostream>
using namespace std;

float calcPCT(float total, float lostpoints)
{
  float pct;
  pct = (total - lostpoints)/total;
  return pct;
}


int main()
{
//Declare one or more variables that will eventually store the input
 float QuizTotal;
 float WrongAnswers;
 float pct; 

// Use cout to display instructions to the user
 cout << "Enter total possible points for the Quiz: ";
// Use cin to get the user's input and store it in the variables you declared previously
 cin >> QuizTotal;

// Use cout to display instructions to the user
 cout << "Enter points for the questions you missed: ";
// Use cin to get the user's input and store it in the variables you declared previously
 cin >> WrongAnswers;

// pass the input along to the calcPCT function by including the variable(s) as arguments 
 pct = calcPCT(QuizTotal, WrongAnswers);
 cout << pct << endl;
}
When you run the program, the program will pause when it gets to the cin instruction. Type your input directly in the output window and hit enter. Here is what it might look like if you entered the values 100 and 10.

calcPCTwithInput.png

Run the program several times with different input for testing.

Note that I used different variable names for the input in main than I used in the calcPCT parameter list. This is just a style choice. In earlier topics, we used the same names for both. Either approach is fine. The variable names are not passed to the function, only the values that are stored in those variables are passed to the function, so the variable names do not matter at all.