Create strings to chars
tring Conversions
Converting Between Strings and Chars
strings to chars
Going from strings and chars is very easy.  Since strings are made from chars, all we have to do is specify the position in the string that we want to access.  To grab a single character from a string we can retrieve that char from the string and store it into a separate char variable.  In the other direction, if we want to put a single character into a string we can assign a new value into that position of the string.  Either way, we have to be able to specify the position of a particular character in a long string. We use bracket notation str[p] to access the pth position in string str.  Counting always begins with 0, so that str[0] is the first letter in str, and if string has length n, then the last character in the string is str[n-1].

Here is a simple example:

#include <iostream>
#include <string>
using namespace std;
 
int main()
{ 
   string myString = "30 Year Adjustable";
   char let1, let2;

   let1 = myString[0];
   let2 = myString[3];

   cout << "Original string: \n"
      << myString << endl;
   cout << "Letters in position 0 and 3: \n"
      << let1 << " " << let2 << " " << endl;

   // now make changes to some characters in the string
   myString[0] = '1';   
   myString[1] = '5';   

   cout << "Modified string: \n"
      << myString << endl;
}
The output is:

Modified string: 15 Year Adjustable
Again, let me emphasize, the numbering of the positions begins with 0.  This will be true for arrays when we get to them.