Create chars to strings
why does  myStr[2] output "P" in :  #include <iostream>
#include <string>
using namespace std;
int main()
{
 string myStr;
  char myChar = 'y';
  char myCharArray[2];

  // convert char to a CString
  myCharArray[0] = myChar;
  myCharArray[1] = '\0';
  
  myStr = string(myCharArray);
  cout << myStr << endl;
  cout << myStr[2] << endl;
  cout << myStr[1];}
View other drafts




























The output of the code you provided will be:

y
P
y
The reason that myStr[2] outputs "P" is because of the way that the string class is implemented in C++. When you create a string object from a C-string, the C-string is copied into the string object. The C-string is terminated by a null character (\0), which is also copied into the string object. This means that the string object will always have a null character at the end, even if the C-string it was created from did not have one.

In your code, you create a string object from the C-string myCharArray. The C-string myCharArray contains the character 'y' and a null character. When the string object is created, it will contain the characters 'y' and '\0'. This means that the character at index 2 of the string object will be 'P', which is the ASCII code for the null character.

Here is a diagram of the string object after it is created:

| y | \0 | P | \0 | ... |
The character at index 2 is highlighted in red.

It is important to note that the string class is not designed to be accessed directly using indices. Instead, you should use the methods provided by the string class to access the characters in the string. For example, you could use the at() method to access the character at index 2 of the string object:

char myChar = myStr.at(2);
The at() method will throw an exception if the index is out of bounds.

Another way to access the characters in a string object is to use a range-based for loop:

for (char c : myStr) {
  // do something with the character c
}
This will iterate over all of the characters in the string object, in order.

It is generally best to avoid accessing string objects directly using indices, as this can lead to unexpected results.